+++ src/lib/utils/http-param-encoder.ts
@@ -1,0 +1,19 @@
+import {
+    HttpParameterCodec,
+    HttpUrlEncodingCodec
+} from '@angular/common/http';
+
+export class QueryEncoder extends HttpUrlEncodingCodec
+{
+    public encodeKey(key:string):string
+    {
+        return encodeURIComponent(key);
+    }
+
+    public encodeValue(value:string):string
+    {
+        return encodeURIComponent(value);
+    }
+}
+
+export const terraHttpParamEncoder:HttpParameterCodec = new QueryEncoder();
